requirements.txt
Django
django-extensions
bpython
django-chartjs
